Add diff and norm to SmbXcorr
Cross-correlations should be applied to differenced values, by default, to remove the effect of a trend in the system as a whole.  Set default behavior of `SmbXcorr` to difference and to accept 0 differencing if desired. For intercomparability, arrays should also be normalized before their cross-correlation is computed. Add this as default option.