Create catchment-integrate-smb.py
Sum the surface mass balance (from HIRHAM) over the Helheim surface catchment, to make a catchment-wide single time series for analysis.

Note that this requires packages `pyshp` and `shapely`, which do not play well with some of the other `nifl` environment packages.  Set up a separate environment to perform this procesisng.