Test xcorr between long-term signals
Define a new function based on `Xcorr1D` that chooses only the long-term-varying components of the velocity signal (eventually incorporate this as an optional argument to the original function in nifl).  Fit a univariate spline to the terminus position record and test the cross-correlation between the long-term portions of the two signals.